add company and user endpoints

+++ opportunity_management_system/src/main/java/com/atc/opportunity_management_system/entity/DealStage.java
@@ -22,7 +22,7 @@ import lombok.Data;
 public class DealStage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer dealStageId;
+    private Long dealStageId;
 
     @Column(name = "dealStageName", nullable = false)
     @NotEmpty
